реализация деления на модули
